refactor(mastodon-api): Die Web-Push-Krypto gehoert nicht dem Adapter (v7.531.8)
VAPID (RFC 8292) und aes128gcm (RFC 8291) lagen im Mastodon-Adapter, wo
eine Subscription an einem API-Token haengt. Die installierte App hat
keinen Token und kaeme an nichts davon heran, ohne das
MASTODON_API_ENABLED-Tor mitzuschleppen -- ein Tor, das die eigene App
eines Mitglieds gerade nicht stummschalten darf.

Vutuv.WebPush besitzt jetzt Schluessel und Zustellung,
Vutuv.MastodonApi.WebPush ist ein duenner Delegate mit seinem eigenen
Tor. Die Changeset-Regeln ziehen nach Vutuv.WebPush.Validations, damit
das naechste Schema den SSRF-Check nicht still verliert. push/2 besitzt
das Loeschen einer Subscription, die der Dienst als tot meldet -- das
Versprechen, das der Dispatcher bisher von Hand gehalten hat.

Verhalten bleibt gleich: keine neue Tabelle, keine Route, keine Config.
Erste Haelfte von #1729, auf Stefans Wunsch aus #1752 herausgeloest.
